WIP: styled updated, tests to fix
