+++ src/components/LinkTo/LinkTo.styles.ts
@@ -1,0 +1,44 @@
+import styled from "styled-components";
+import resetCSS from "../../styles/reset";
+import color from "../../styles/colors";
+import fonts from "../../styles/fonts";
+
+// types
+import type { LinkToProps } from "./types";
+
+// styles
+type TStyleProps = Pick<LinkToProps, "iconLayout">;
+
+const LinkStyled = styled.a`
+	${resetCSS}
+	${fonts.text}
+  align-items: center;
+	color: ${color.blue};
+	display: inline-block;
+	width: fit-content;
+
+	&:hover {
+		filter: brightness(0.7);
+	}
+`;
+
+const FlexSpanStyled = styled.span<TStyleProps>`
+	align-items: baseline;
+	display: flex;
+	flex-direction: ${({ iconLayout }) =>
+		iconLayout === "leading" ? "row" : "row-reverse"};
+	margin: ${(p) =>
+		p.iconLayout === "leading" ? "0 2px 0 4px" : "0 4px 0 2px"};
+
+	svg {
+		margin: ${(p) =>
+			p.iconLayout === "leading" ? "0 4px 0 0" : "0 0 0 4px"};
+	}
+`;
+
+const LinkToStyles = {
+	LinkStyled,
+	FlexSpanStyled,
+};
+
+export default LinkToStyles;
